refactored environment empty section
